+++ components/collector/tests/source_collectors/axe_core/test_accessibility.py
@@ -117,6 +117,18 @@ class AxeCoreAccessibilityTest(AxeCoreTestCase):
         response = await self.collect(get_request_json_return_value=self.json)
         self.assert_measurement(response, value="1", entities=[self.expected_entities[1]])
 
+    async def test_element_include_filter(self):
+        """Test that violations can be filtered by element."""
+        self.set_source_parameter("element_include_filter", ["html1"])
+        response = await self.collect(get_request_json_return_value=self.json)
+        self.assert_measurement(response, value="1", entities=[self.expected_entities[0]])
+
+    async def test_element_exclude_filter(self):
+        """Test that violations can be filtered by element."""
+        self.set_source_parameter("element_exclude_filter", ["html1"])
+        response = await self.collect(get_request_json_return_value=self.json)
+        self.assert_measurement(response, value="1", entities=[self.expected_entities[1]])
+
     async def test_zipped_json(self):
         """Test that a zip archive with JSON files is processed correctly."""
         self.set_source_parameter("url", "axe.zip")